backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
cryptography==42.0.0

# AI & NLP
openai==1.10.0

# Communication
twilio==8.11.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Date/Time
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# CORS
python-cors==1.0.0

# Translation/i18n
googletrans==4.0.0rc1

# PDF Generation
reportlab==4.0.8
